requirements.txt
pyqt6~=6.3.0
numpy~=1.22.4
opencv-python~=4.5.5.64
pyexiv2~=2.7.1
pyyaml~=6.0
mysql-connector-python~=8.0.29